requirements.txt
streamlit
pandas
scikit-learn
scipy
scikit-surprise
